refactor: move all in game plugins under ingame module.
