+++ src/in_game.rs
@@ -1,0 +1,7 @@
+pub mod destroy_enemy;
+pub mod enemy;
+pub mod life_count;
+pub mod play_area;
+pub mod player;
+pub mod player_shot;
+pub mod scoreboard;
