+++ src/in_game/enemy/assets_holder.rs
