+++ src/in_game/enemy/barrage/configuration.rs
