+++ src/in_game/enemy/barrage/plugin.rs
@@ -1,14 +1,14 @@
 use super::bullet::Bullet;
 use crate::app_state::AppState;
-use crate::enemy::barrage::bullet::BulletType;
-use crate::enemy::barrage::bulletml_runner::BulletMLRunner;
-use crate::enemy::barrage::bulletml_runner::BulletMLRunnerData;
-use crate::enemy::barrage::configuration::BarrageConfiguration;
-use crate::enemy::system_label::EnemySystemLabel;
-use crate::enemy::Enemy;
-use crate::life_count::LifeCount;
-use crate::play_area::PlayAreaDescriptor;
-use crate::player::Player;
+use crate::in_game::enemy::barrage::bullet::BulletType;
+use crate::in_game::enemy::barrage::bulletml_runner::BulletMLRunner;
+use crate::in_game::enemy::barrage::bulletml_runner::BulletMLRunnerData;
+use crate::in_game::enemy::barrage::configuration::BarrageConfiguration;
+use crate::in_game::enemy::system_label::EnemySystemLabel;
+use crate::in_game::enemy::Enemy;
+use crate::in_game::life_count::LifeCount;
+use crate::in_game::play_area::PlayAreaDescriptor;
+use crate::in_game::player::Player;
 use bevy::prelude::*;
 use bevy_bulletml::BulletMLServer;
 use bevy_bulletml::Runner;
