+++ src/in_game/enemy/emerge.rs
@@ -1,11 +1,11 @@
 use crate::app_state::AppState;
-use crate::enemy::assets_holder::EnemyAssetsHolder;
-use crate::enemy::barrage::configuration::BarrageConfiguration;
-use crate::enemy::movement::move_pattern::MovePattern;
-use crate::enemy::system_label::EnemySystemLabel;
-use crate::enemy::Enemy;
-use crate::life_count::LifeCount;
-use crate::play_area::PlayAreaDescriptor;
+use crate::in_game::enemy::assets_holder::EnemyAssetsHolder;
+use crate::in_game::enemy::barrage::configuration::BarrageConfiguration;
+use crate::in_game::enemy::movement::move_pattern::MovePattern;
+use crate::in_game::enemy::system_label::EnemySystemLabel;
+use crate::in_game::enemy::Enemy;
+use crate::in_game::life_count::LifeCount;
+use crate::in_game::play_area::PlayAreaDescriptor;
 use bevy::prelude::*;
 use std::collections::HashMap;
 use std::fs::File;
