+++ src/in_game/enemy/movement.rs
@@ -1,8 +1,8 @@
 pub mod move_pattern;
 use super::system_label::EnemySystemLabel;
 use crate::app_state::AppState;
-use crate::enemy::Enemy;
-use crate::life_count::LifeCount;
+use crate::in_game::enemy::Enemy;
+use crate::in_game::life_count::LifeCount;
 use bevy::prelude::*;
 use move_pattern::MovePattern;
