+++ src/in_game/life_count.rs
