+++ src/in_game/play_area/play_area_descriptor.rs
